Move Duration to domain folder and add toSeconds method
- Move Duration.ts from utils/ to domain/ as it's a domain concept
- Add toSeconds() method to Duration class for WorklogEntry integration
- Update all Duration imports across codebase
- Maintains consistency with domain-driven design approach
