+++ source/components/AttendanceEditForm.tsx
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {Box, Text, useInput, useFocus} from 'ink';
-import {Duration} from '../utils/Duration.js';
+import {Duration} from '../domain/Duration.js';
 import type {Attendance} from '../attendance/types.js';
 import TimeInputField from './TimeInputField.js';
 import DurationInput from './WorklogForm/DurationInput.js';
