+++ source/components/areas/WorklogFormArea.tsx
@@ -3,7 +3,7 @@ import {Box} from 'ink';
 import {InlineWorklogForm} from '../InlineWorklogForm.js';
 import type {WorklogFormData} from '../../hooks/useWorklogForm.js';
 import type {JiraConfig, WorklogEntry, JiraClient} from '../../jira-client.js';
-import type {Duration} from '../../utils/Duration.js';
+import type {Duration} from '../../domain/Duration.js';
 
 export type WorklogFormAreaProps = {
 	worklogForm: WorklogFormData;
