+++ source/hooks/useTitleResolver.test.ts
@@ -1,5 +1,5 @@
 import test from 'ava';
-import {Duration} from '../utils/Duration.js';
+import {Duration} from '../domain/Duration.js';
 import {useTitleResolver} from './useTitleResolver.js';
 import type {
 	DeleteCandidate,
